Lock scrolls to 0.3.3
This is the version that was being used prior to the ruby upgrade. With
that upgrade, I bumped scrolls to 0.3.9. I noticed in staging that the
logs were being flooded with deprication warnings. Additionally Scrolls
doesn't support the "-W0" ruby flag for silencing deprication warnings.
Given the volume of these messages, I chose to down-grade Scrolls to the
earlier released version to resolve this for now.